requirements.txt
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-anthropic>=0.1.0
pydantic>=2.0.0
pypdf>=3.0.0
python-docx>=1.0.0
streamlit>=1.28.0
